phase-4-kubernetes/backend/requirements.txt
# Web Framework
fastapi==0.128.0
uvicorn[standard]==0.27.0

# Database
sqlmodel==0.0.14
asyncpg==0.29.0
alembic==1.13.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx<0.28.0  # Pin to <0.28.0 for openai 1.54.0 compatibility

# Development
ruff==0.1.15
mypy==1.8.0

# Phase III - AI Chatbot
openai==1.54.0
mcp==1.1.2
python-dotenv==1.0.0  # For loading .env files
